refactor(signals): rename with-sort folder to with-feature-factory
