+++ libs/ngrx-traits/signals/src/lib/with-call-status/with-call-status.ts
@@ -9,12 +9,12 @@ import {
   withState,
 } from '@ngrx/signals';
 
-import { withFeatureFactory } from '../with-store/with-feature-factory';
+import { withFeatureFactory } from '../with-feature-factory/with-feature-factory';
 import {
   FeatureConfigFactory,
   getFeatureConfig,
   StoreSource,
-} from '../with-store/with-feature-factory.model';
+} from '../with-feature-factory/with-feature-factory.model';
 import {
   CallStatus,
   CallStatusComputed,
