+++ libs/ngrx-traits/signals/src/lib/with-entities-filter/with-entities-local-filter.ts
@@ -27,12 +27,12 @@ import {
   broadcast,
   withEventHandler,
 } from '../with-event-handler/with-event-handler';
-import { withFeatureFactory } from '../with-store/with-feature-factory';
+import { withFeatureFactory } from '../with-feature-factory/with-feature-factory';
 import {
   FeatureConfigFactory,
   getFeatureConfig,
   StoreSource,
-} from '../with-store/with-feature-factory.model';
+} from '../with-feature-factory/with-feature-factory.model';
 import {
   debounceFilterPipe,
   getWithEntitiesFilterEvents,
