+++ libs/ngrx-traits/signals/src/lib/with-entities-selection/with-entities-multi-selection.ts
@@ -25,12 +25,12 @@ import {
   onEvent,
   withEventHandler,
 } from '../with-event-handler/with-event-handler';
-import { withFeatureFactory } from '../with-store/with-feature-factory';
+import { withFeatureFactory } from '../with-feature-factory/with-feature-factory';
 import {
   FeatureConfigFactory,
   getFeatureConfig,
   StoreSource,
-} from '../with-store/with-feature-factory.model';
+} from '../with-feature-factory/with-feature-factory.model';
 import {
   EntitiesMultiSelectionComputed,
   EntitiesMultiSelectionMethods,
